feat: 문제관련 예외처리 구현

+++ src/main/java/algo_arena/problem/service/ProblemService.java
@@ -1,8 +1,12 @@
 package algo_arena.problem.service;
 
+import static algo_arena.common.exception.enums.ErrorType.*;
+
 import algo_arena.member.entity.Member;
-import algo_arena.member.repository.MemberRepository;
+import algo_arena.member.exception.AuthException;
+import algo_arena.member.service.MemberService;
 import algo_arena.problem.entity.Problem;
+import algo_arena.problem.exception.ProblemException;
 import algo_arena.problem.repository.ProblemRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
@@ -16,27 +20,34 @@ import org.springframework.util.StringUtils;
 public class ProblemService {
 
     private final ProblemRepository problemRepository;
-    private final MemberRepository memberRepository;
+    private final MemberService memberService;
 
     @Transactional
     public Problem registerProblem(Problem problem, String memberName) {
-        checkAuth(memberName);
+        checkAdminAuth(memberName);
         return problemRepository.save(problem);
     }
 
     public List<Problem> findProblemsBySearch(Long minNumber, Long maxNumber, String title) {
+        if (minNumber != null && maxNumber != null && minNumber > maxNumber) {
+            throw new ProblemException(INVALID_PROBLEM_NUMBER_RANGE);
+        }
         return problemRepository.findProblemsBySearch(minNumber, maxNumber, title);
     }
 
-    public Problem findProblemByNumber(Long id) {
-        return problemRepository.findByNumber(id).orElseThrow();
+    public Problem findProblemByNumber(Long number) {
+        if (number == null) {
+            throw new ProblemException(NULL_VALUE);
+        }
+        return problemRepository.findByNumber(number)
+            .orElseThrow(() -> new ProblemException(PROBLEM_NOT_FOUND));
     }
 
     @Transactional
     public void updateProblem(Long number, String title, String link, String memberName) {
-        checkAuth(memberName);
-
+        checkAdminAuth(memberName);
         Problem problem = findProblemByNumber(number);
+
         if (StringUtils.hasText(title)) {
             problem.changeTitle(title);
         }
@@ -47,14 +58,17 @@ public class ProblemService {
 
     @Transactional
     public void deleteProblem(Long number, String memberName) {
-        checkAuth(memberName);
+        checkAdminAuth(memberName);
+        if (number == null) {
+            throw new ProblemException(NULL_VALUE);
+        }
         problemRepository.deleteByNumber(number);
     }
 
-    private void checkAuth(String memberName) {
-        Member member = memberRepository.findByName(memberName).orElseThrow();
+    private void checkAdminAuth(String memberName) {
+        Member member = memberService.findMemberByName(memberName);
         if (!member.isAdmin()) {
-            throw new RuntimeException("권한이 없습니다. 관리자에게 문의하세요.");
+            throw new AuthException(INVALID_ROLE);
         }
     }
 }
